+++ packages/db/src/index.ts
@@ -6,4 +6,4 @@ export {
   workspaceSchema,
   systemSchema,
 } from "./schema/index.js";
-export * from "./create-env.js";
+export * from "./upsert-env.js";
